chore(nft-studio): separate base form inputs and formik
