--- nft-studio/src/components/form/formik/ErrorMessage.tsx
@@ -1,6 +1,0 @@
-import { ErrorMessage as FormikErrorMessage } from 'formik'
-import styled from 'styled-components'
-
-export const ErrorMessage = styled(FormikErrorMessage)`
-  color: red;
-`
