+++ nft-studio/src/components/form/formik/RadioButton.tsx
@@ -1,11 +1,6 @@
-import { IconCheckCircle, IconCircle, Shelf, Text } from '@centrifuge/fabric'
 import { Field, useField } from 'formik'
 import React from 'react'
-import styled from 'styled-components'
-
-interface RadioInputProps extends React.InputHTMLAttributes<HTMLInputElement> {
-  name: string
-}
+import { RadioButton as RadioButtonBase } from '../base/RadioButton'
 
 interface RadioButtonProps extends React.InputHTMLAttributes<HTMLInputElement> {
   label: string
@@ -13,35 +8,7 @@ interface RadioButtonProps extends React.InputHTMLAttributes<HTMLInputElement> {
   value: string
 }
 
-const HiddenField = styled(Field)`
-  display: none;
-`
-
-const FocusableLabel = styled.label`
-  :focus {
-    color: ${({ theme }) => theme.colors.brand};
-  }
-`
-
-export const RadioInput: React.FC<RadioInputProps> = (props) => {
-  const [field] = useField(props.name)
-
-  return (
-    <label>
-      {field.value === props.value ? <IconCheckCircle /> : <IconCircle />}
-      <HiddenField type="radio" tabIndex={-1} {...field} {...props} />
-    </label>
-  )
-}
-
-export const RadioButton: React.FC<RadioButtonProps> = ({ label, ...props }) => {
-  const id = props.id || props.name
-  return (
-    <FocusableLabel tabIndex={0} htmlFor={id}>
-      <Shelf gap="1">
-        <RadioInput type="radio" {...props} id={id} />
-        <Text variant="interactive2">{label}</Text>
-      </Shelf>
-    </FocusableLabel>
-  )
+export const RadioButton: React.FC<RadioButtonProps> = (props) => {
+  const [field] = useField<string>(props)
+  return <Field as={RadioButtonBase} {...props} checked={field.value === props.value} />
 }
